Moved bracket page to it's own folder
